// src/frontend/web/src/gen/quadsmith/media_pb.ts
// @generated by protoc-gen-es v2.16.0 with parameter "target=ts"
// @generated from file media.proto (package quadsmith, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file media.proto.
 */
export const file_media: GenFile =
  /*@__PURE__*/
  fileDesc(
    "CgttZWRpYS5wcm90bxIJcXVhZHNtaXRoIoABCgVNZWRpYRIiCgR0eXBlGAEgASgOMhQucXVhZHNtaXRoLk1lZGlhVHlwZRILCgN1cmwYAiABKAkSEgoFdGl0bGUYAyABKAlIAIgBARIYCgtkZXNjcmlwdGlvbhgEIAEoCUgBiAEBQggKBl90aXRsZUIOCgxfZGVzY3JpcHRpb24qVQoJTWVkaWFUeXBlEhoKFk1FRElBX1RZUEVfVU5TUEVDSUZJRUQQABIUChBNRURJQV9UWVBFX0lNQUdFEAESFgoSTUVESUFfVFlQRV9ZT1VUVUJFEAJCHVobcXVhZHNtaXRoL2FwaS9nZW4vcXVhZHNtaXRoYgZwcm90bzM",
  );

/**
 * @generated from message quadsmith.Media
 */
export type Media = Message<"quadsmith.Media"> & {
  /**
   * @generated from field: quadsmith.MediaType type = 1;
   */
  type: MediaType;

  /**
   * @generated from field: string url = 2;
   */
  url: string;

  /**
   * @generated from field: optional string title = 3;
   */
  title?: string | undefined;

  /**
   * @generated from field: optional string description = 4;
   */
  description?: string | undefined;
};

/**
 * Describes the message quadsmith.Media.
 * Use `create(MediaSchema)` to create a new message.
 */
export const MediaSchema: GenMessage<Media> = /*@__PURE__*/ messageDesc(file_media, 0);

/**
 * @generated from enum quadsmith.MediaType
 */
export enum MediaType {
  /**
   * @generated from enum value: MEDIA_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDIA_TYPE_IMAGE = 1;
   */
  IMAGE = 1,

  /**
   * @generated from enum value: MEDIA_TYPE_YOUTUBE = 2;
   */
  YOUTUBE = 2,
}

/**
 * Describes the enum quadsmith.MediaType.
 */
export const MediaTypeSchema: GenEnum<MediaType> = /*@__PURE__*/ enumDesc(file_media, 0);